introduce separation between users and games
This change introduces two new executables/processes: gamelet and
central. The gamelet is an internal API for a game. The central is an
external API for user management and joining games. It might help to see
the gamelet, api, updater and worker as a game server, while central is
the main server for listing and joining game servers.

The central uses a sqlite db while the games are still heavily using redis.