central: change join game operation order
When joining a game the central need to both store that the user joined
the game, but also notify the gamelet of this (so that the user is
created in that game). Previously this was done by first storing the
central record and then notifying the gamelet. This did not work well in
cases when the gamelet failed to join the user, because the central
would still have persisted the record that the user joined.

This commit changes so that the operations happen the other way around.
First the gamelet is notified, then the central record is stored. This
will probably still not work if storing the central record would fail,
but that risk is much lower.